Make read_json_value handle "true" and "false"
When I was attempting to push my docs back to CouchDB, I found that
some of my design documents had certain properties serialized as
literally "false" or "true". I don't necessarily disagree with that,
but then the JSON parser complains. So this makes it return false or
true when that's all in the file, which makes my docs import as
expected.

If there's a cleaner way to fix this, please let me know and I'll
submit a better patch :)